move userinfo schema extension validation to utils
